-- updated as per Brett comment: "Adrian, generally like the patch but I'm concerned about the terminology.
The term 'mute' whilst somewhat applicable here, is also used to 'mute'
a channel (i.e. set the channel so that one party can't hear) which has
nothing to recording. I'm wondering of we should be more direct and
create a 'recording' exception that is thrown for any circumstance when
there is a problem starting/stopping/pausing/resuming recordings? "