Bring back ConcurrentHashMap and properly synchronize
Because the list of active bridges can be accessed from multiple
threads, it does make more sense to use a ConcurrentHashMap, so bring
it back. Also fix a CAS in BridgesActive.createBridge() and make
activeBridges private so there are no shenanigans (it was already
package-private, so this isn't that significant).

For BridgeState, stick with a HashMap and use proper synchronization.
This is necessary to synchronize the put/remove/size semantics present
in addMember/removeMember.