Use ParkingSpace to set parkedAt on channel
Use ParkingSpace to set parkedAt on channel instead of 'Exten'. Falls back to
Exten if ParkingSpace is null for some reason (most likely because we are on
an older version of Asterisk).

Also, changes setParkedAt to include a parkinglot name. This can be retrieved
from the Asterisk channel with getParkingLot(). Between getParkedAt() and
getParkingLot() you can get all the arguments needed to use the ParkedCall
dialplan application on any parked call.