ManagerReaderImpl.java: Remove `RTCPSent`/`RTCPReceived` parsing workaround.
Both the incorrect `RTCPSent` and `RTCPReceived` events and the commit that fixed them were included in Asterisk 1.6.0.

There was never a release that included the broken events so there's no reason to continue working around it. If you're running a beta of Asterisk 1.6.0 that contains this bug, you should consider upgrading.